requirements.txt
```
openai
faiss-cpu
```